Split eip_7251_root.go into different files and reuse/share code with historical state summaries root. It's identical!
