+++ beacon-chain/state/stateutil/BUILD.bazel
@@ -4,7 +4,6 @@ go_library(
     name = "go_default_library",
     srcs = [
         "block_header_root.go",
-        "eip_7251_root.go",
         "eth1_root.go",
         "field_root_attestation.go",
         "field_root_eth1.go",
@@ -13,7 +12,11 @@ go_library(
         "historical_summaries_root.go",
         "participation_bit_root.go",
         "pending_attestation_root.go",
+        "pending_balance_deposits_root.go",
+        "pending_consolidations_root.go",
+        "pending_partial_withdrawals_root.go",
         "reference.go",
+        "slice_root.go",
         "sync_committee.root.go",
         "trie_helpers.go",
         "unrealized_justification.go",
