+++ beacon-chain/state/stateutil/pending_partial_withdrawals_root.go
@@ -5,6 +5,6 @@ import (
 	ethpb "github.com/prysmaticlabs/prysm/v5/proto/prysm/v1alpha1"
 )
 
-func HistoricalSummariesRoot(summaries []*ethpb.HistoricalSummary) ([32]byte, error) {
-	return SliceRoot(summaries, fieldparams.HistoricalRootsLength)
+func PendingPartialWithdrawalsRoot(slice []*ethpb.PendingPartialWithdrawal) ([32]byte, error) {
+	return SliceRoot(slice, fieldparams.PendingPartialWithdrawalsLimit)
 }
